requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pandas==2.2.1
pydantic==2.10.3
python-multipart==0.0.9
pypdf==4.1.0
weasyprint==62.1
scikit-learn==1.4.1.post1
numpy==1.26.4
jinja2==3.1.3
google-generativeai==0.3.2
beautifulsoup4==4.12.3
requests==2.31.0
aiohttp==3.9.3
python-dotenv==1.0.1
